refactor: move Autoformer layers to shared layers module and address review feedback
- Move AutoCorrelation, AutoCorrelationLayer to layers/_attention/
- Move SeasonalLayerNorm (renamed from my_Layernorm) to layers/_normalization/
- Move AutoformerEncoder/EncoderLayer to layers/_encoders/
- Move AutoformerDecoder/DecoderLayer to layers/_decoders/ (new)
- Move Autoformer embeddings to layers/_embeddings/
- Reuse existing SeriesDecomposition and MovingAvg from shared layers
- Convert sub_modules.py to thin re-export shim for backward compat
- Remove unnecessary _get_test_datamodule_from from Autoformer_pkg_v2
