+++ pytorch_forecasting/layers/_attention/__init__.py
@@ -3,9 +3,19 @@ Attention Layers for pytorch-forecasting models.
 """
 
 from pytorch_forecasting.layers._attention._attention_layer import AttentionLayer
+from pytorch_forecasting.layers._attention._auto_correlation import (
+    AutoCorrelation,
+    AutoCorrelationLayer,
+)
 from pytorch_forecasting.layers._attention._full_attention import (
     FullAttention,
     TriangularCausalMask,
 )
 
-__all__ = ["AttentionLayer", "FullAttention", "TriangularCausalMask"]
+__all__ = [
+    "AttentionLayer",
+    "AutoCorrelation",
+    "AutoCorrelationLayer",
+    "FullAttention",
+    "TriangularCausalMask",
+]
