+++ pytorch_forecasting/layers/_embeddings/__init__.py
@@ -2,6 +2,13 @@
 Implementation of embedding layers for PTF models imported from `nn.Modules`
 """
 
+from pytorch_forecasting.layers._embeddings._autoformer_embedding import (
+    DataEmbedding_wo_pos,
+    FixedEmbedding,
+    TemporalEmbedding,
+    TimeFeatureEmbedding,
+    TokenEmbedding,
+)
 from pytorch_forecasting.layers._embeddings._data_embedding import (
     DataEmbedding_inverted,
 )
@@ -12,8 +19,13 @@ from pytorch_forecasting.layers._embeddings._positional_embedding import (
 from pytorch_forecasting.layers._embeddings._sub_nn import embedding_cat_variables
 
 __all__ = [
-    "PositionalEmbedding",
+    "DataEmbedding_wo_pos",
     "DataEmbedding_inverted",
     "EnEmbedding",
+    "FixedEmbedding",
+    "PositionalEmbedding",
+    "TemporalEmbedding",
+    "TimeFeatureEmbedding",
+    "TokenEmbedding",
     "embedding_cat_variables",
 ]
