+++ pytorch_forecasting/layers/_encoders/__init__.py
@@ -2,7 +2,11 @@
 Encoder layers for neural network models.
 """
 
+from pytorch_forecasting.layers._encoders._autoformer_encoder import (
+    AutoformerEncoder,
+    AutoformerEncoderLayer,
+)
 from pytorch_forecasting.layers._encoders._encoder import Encoder
 from pytorch_forecasting.layers._encoders._encoder_layer import EncoderLayer
 
-__all__ = ["Encoder", "EncoderLayer"]
+__all__ = ["AutoformerEncoder", "AutoformerEncoderLayer", "Encoder", "EncoderLayer"]
