+++ pytorch_forecasting/layers/_encoders/_autoformer_encoder.py
@@ -1,0 +1,86 @@
+"""
+Autoformer encoder with progressive decomposition architecture.
+
+Implements encoder layers that incorporate series decomposition after each
+attention and feed-forward sub-layer, progressively separating trend from
+seasonal components.
+"""
+
+import torch.nn as nn
+import torch.nn.functional as F
+
+from pytorch_forecasting.layers._decomposition._series_decomp import (
+    SeriesDecomposition,
+)
+
+
+class AutoformerEncoderLayer(nn.Module):
+    """
+    Autoformer encoder layer with the progressive decomposition architecture.
+    """
+
+    def __init__(
+        self,
+        attention,
+        d_model,
+        d_ff=None,
+        moving_avg=25,
+        dropout=0.1,
+        activation="relu",
+    ):
+        super().__init__()
+        d_ff = d_ff or 4 * d_model
+        self.attention = attention
+        self.conv1 = nn.Conv1d(
+            in_channels=d_model, out_channels=d_ff, kernel_size=1, bias=False
+        )
+        self.conv2 = nn.Conv1d(
+            in_channels=d_ff, out_channels=d_model, kernel_size=1, bias=False
+        )
+        self.decomp1 = SeriesDecomposition(moving_avg)
+        self.decomp2 = SeriesDecomposition(moving_avg)
+        self.dropout = nn.Dropout(dropout)
+        self.activation = F.relu if activation == "relu" else F.gelu
+
+    def forward(self, x, attn_mask=None):
+        new_x, attn = self.attention(x, x, x, attn_mask=attn_mask)
+        x = x + self.dropout(new_x)
+        x, _ = self.decomp1(x)
+        y = x
+        y = self.dropout(self.activation(self.conv1(y.transpose(-1, 1))))
+        y = self.dropout(self.conv2(y).transpose(-1, 1))
+        res, _ = self.decomp2(x + y)
+        return res, attn
+
+
+class AutoformerEncoder(nn.Module):
+    """
+    Autoformer encoder.
+    """
+
+    def __init__(self, attn_layers, conv_layers=None, norm_layer=None):
+        super().__init__()
+        self.attn_layers = nn.ModuleList(attn_layers)
+        self.conv_layers = (
+            nn.ModuleList(conv_layers) if conv_layers is not None else None
+        )
+        self.norm = norm_layer
+
+    def forward(self, x, attn_mask=None):
+        attns = []
+        if self.conv_layers is not None:
+            for attn_layer, conv_layer in zip(self.attn_layers, self.conv_layers):
+                x, attn = attn_layer(x, attn_mask=attn_mask)
+                x = conv_layer(x)
+                attns.append(attn)
+            x, attn = self.attn_layers[-1](x)
+            attns.append(attn)
+        else:
+            for attn_layer in self.attn_layers:
+                x, attn = attn_layer(x, attn_mask=attn_mask)
+                attns.append(attn)
+
+        if self.norm is not None:
+            x = self.norm(x)
+
+        return x, attns
