+++ pytorch_forecasting/layers/_normalization/__init__.py
@@ -1,7 +1,10 @@
 """
-RevIN: Reverse Instance Normalization
+Normalization layers for pytorch-forecasting models.
 """
 
 from pytorch_forecasting.layers._normalization._revin import RevIN
+from pytorch_forecasting.layers._normalization._seasonal_layernorm import (
+    SeasonalLayerNorm,
+)
 
-__all__ = ["RevIN"]
+__all__ = ["RevIN", "SeasonalLayerNorm"]
